requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
dacite==1.9.2
faiss-cpu==1.13.0
filelock==3.20.0
fsspec==2025.10.0
ftfy==6.3.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
regex==2025.11.3
requests==2.32.5
sympy==1.14.0
tcgdex-sdk==2.2.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
